refactor(executors): 按运行时归并目录
