+++ src/executors/api/openai.ts
@@ -1,8 +1,8 @@
-import type { ExecResult, ExecutorInput } from '../types/index.js';
-import { optionalTokenCount, splitInclusiveInputTokens } from '../shared/token-usage.js';
-import { DEFAULT_TIMEOUT_MS } from './defaults.js';
+import type { ExecResult, ExecutorInput } from '../../types/index.js';
+import { optionalTokenCount, splitInclusiveInputTokens } from '../../shared/token-usage.js';
+import { DEFAULT_TIMEOUT_MS } from '../defaults.js';
 import { readJsonResponse, responseBodyPreview } from './http.js';
-import { asErrorLike, errorMessage } from './runtime.js';
+import { asErrorLike, errorMessage } from '../runtime.js';
 
 interface OpenAiResponse {
   usage?: {
