+++ src/executors/claude/sdk-trace.ts
@@ -1,8 +1,8 @@
-import type { ToolCallInfo, TurnInfo } from '../types/index.js';
-import type { ClaudeSdkBaseMessage } from './claude-protocol.js';
-import { safeSliceForJson } from '../util/safe-slice.js';
-import { isToolResultFailureText } from '../observability/text-signals.js';
-import { normalizeToolIdentity } from '../shared/tool-identity.js';
+import type { ToolCallInfo, TurnInfo } from '../../types/index.js';
+import type { ClaudeSdkBaseMessage } from './protocol.js';
+import { safeSliceForJson } from '../../util/safe-slice.js';
+import { isToolResultFailureText } from '../../observability/text-signals.js';
+import { normalizeToolIdentity } from '../../shared/tool-identity.js';
 
 export function isClaudeSdkResultMessage(message: ClaudeSdkBaseMessage): boolean {
   return message.type === 'result';
