+++ src/executors/index.ts
@@ -1,13 +1,13 @@
 import type { ExecutorFn } from '../types/index.js';
-import { anthropicApiExecutor } from './anthropic-api.js';
-import { claudeCliExecutor } from './claude-cli.js';
-import { claudeSdkExecutor } from './claude-sdk.js';
-import { extractAgentTrace } from './claude-sdk-trace.js';
-import { codexCliExecutor } from './codex-cli.js';
-import { codexSdkExecutor } from './codex-sdk.js';
+import { anthropicApiExecutor } from './api/anthropic.js';
+import { openAiApiExecutor } from './api/openai.js';
+import { claudeCliExecutor } from './claude/cli.js';
+import { claudeSdkExecutor } from './claude/sdk.js';
+import { extractAgentTrace } from './claude/sdk-trace.js';
+import { codexCliExecutor } from './codex/cli.js';
+import { codexSdkExecutor } from './codex/sdk.js';
 import { geminiExecutor } from './gemini.js';
-import { openAiApiExecutor } from './openai-api.js';
-import { createScriptExecutor } from './script.js';
+import { createScriptExecutor } from './script/index.js';
 import { enforceExecutorCapabilities } from './capabilities.js';
 import {
   getExecutorDescriptor,
