+++ src/executors/script/command.ts
