+++ test/executors/codex-protocol.test.ts
@@ -4,8 +4,8 @@ import {
   buildCodexResult,
   extractCodexProtocolError,
   validateCodexProtocol,
-} from '../../src/executors/codex-protocol.js';
-import type { CodexEvent } from '../../src/executors/codex-protocol.js';
+} from '../../src/executors/codex/protocol.js';
+import type { CodexEvent } from '../../src/executors/codex/protocol.js';
 
 describe('Codex protocol normalization', () => {
   it('keeps an item-level error non-fatal when the turn still completes with an answer', () => {
